+++ src/containers/ContactManagement/AdminContactManagement/AdminContactManagement.test.tsx
@@ -5,13 +5,11 @@ import { MockedProvider } from '@apollo/client/testing';
 import { BrowserRouter as Router } from 'react-router-dom';
 
 import { getAllOrganizations } from 'mocks/Organization';
-import { moveContacts } from 'mocks/Contact';
 
 import { setUserSession } from 'services/AuthService';
-import { setNotification } from 'common/notification';
 import { AdminContactManagement } from './AdminContactManagement';
 
-const mocks = [...getAllOrganizations, moveContacts];
+const mocks = getAllOrganizations;
 
 setUserSession(JSON.stringify({ roles: [{ label: 'Admin' }], organization: { id: '1' } }));
 
@@ -43,7 +41,7 @@ test('Admin contact management form renders correctly', async () => {
 test('the page should have a disabled upload button by default', async () => {
   render(contactManagement);
 
-  const uploadButton = await screen.getByTestId('uploadButton');
+  const uploadButton = await screen.getByTestId('moveContactsBtn');
   expect(uploadButton).toBeInTheDocument();
   expect(uploadButton).toHaveAttribute('disabled');
 });
@@ -60,31 +58,5 @@ test('Files other than .csv should raise a warning message upon upload', async (
     expect(
       screen.getByText(/Please make sure the file format matches the sample/)
     ).toBeInTheDocument();
-  });
-});
-
-test('Success Notification should be called upon successful CSV upload', async () => {
-  render(contactManagement);
-
-  // Valid CSV
-  const csvContent = `name,phone,collection
-  John Doe,919876543210,"Optin collection,Optout Collection"
-  Virat Kohli,919876543220,Cricket`;
-  const file = new File([csvContent], 'test.csv', { type: 'text/csv' });
-
-  await waitFor(() => {
-    const fileInput = screen.getByTestId('uploadFile');
-    userEvent.upload(fileInput, file);
-  });
-  await waitFor(() => {
-    // the filename should be visible instead of Select .csv after upload
-    expect(screen.getByText('test.csv')).toBeInTheDocument();
-  });
-
-  const uploadBtn = screen.getByTestId('uploadButton');
-  userEvent.click(uploadBtn);
-
-  await waitFor(() => {
-    expect(setNotification).toHaveBeenCalled();
   });
 });
