+++ src/containers/ContactManagement/UploadContactsDialog/UploadContactsDialog.test.tsx
@@ -1,5 +1,4 @@
-import { render, waitFor, screen, fireEvent } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { render, waitFor } from '@testing-library/react';
 import { MockedProvider } from '@apollo/client/testing';
 
 import { BrowserRouter as Router } from 'react-router-dom';
@@ -9,7 +8,6 @@ import { getAllOrganizations } from 'mocks/Organization';
 import UploadContactsDialog from './UploadContactsDialog';
 import { filterCollectionQuery } from 'mocks/Collection';
 import { CONTACTS_COLLECTION } from 'common/constants';
-import { importContacts } from 'mocks/Contact';
 
 const mocks = [
   ...getAllOrganizations,
@@ -23,7 +21,6 @@ const mocks = [
       order: 'ASC',
     },
   }),
-  importContacts,
 ];
 
 const setDialogMock = vi.fn();
@@ -51,39 +48,4 @@ test('Upload contact dialog renders correctly', async () => {
   await waitFor(() => {
     expect(getByText('Upload Contacts')).toBeInTheDocument();
   });
-});
-
-test('Should be able to upload valid CSV', async () => {
-  render(dialogBox);
-
-  // Valid CSV
-  const csvContent = `name,phone,collection
-  John Doe,919876543210,"Optin collection,Optout Collection"
-  Virat Kohli,919876543220,Cricket`;
-  const file = new File([csvContent], 'test.csv', { type: 'text/csv' });
-
-  await waitFor(() => {
-    const fileInput = screen.getByTestId('import');
-    userEvent.upload(fileInput, file);
-  });
-  await waitFor(() => {
-    // the filename should be visible instead of Select .csv after upload
-    expect(screen.getByText('test.csv')).toBeInTheDocument();
-  });
-
-  const autocomplete = screen.getByTestId('autocomplete-element');
-
-  fireEvent.click(autocomplete);
-  fireEvent.keyDown(autocomplete, { key: 'ArrowDown' });
-  fireEvent.click(screen.getByText('Staff group'));
-
-  fireEvent.click(screen.getByTestId('ok-button'));
-  fireEvent.click(screen.getByText('Are these contacts opted in?'));
-  fireEvent.click(screen.getByTestId('ok-button'));
-
-  await waitFor(() => {
-    expect(screen.getByText('Contact import is in progress.')).toBeInTheDocument();
-  });
-
-  fireEvent.click(screen.getByText('Go to notifications'));
 });
